requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
pywebpush==2.0.1
APScheduler==3.10.4
psycopg2-binary==2.9.10
